test(useUrlState): refactor test
